converted to positive filtering logic
Previous logic excluded all body parts that did not satisfy the thresholding at least once. This meant if (A, B, C) are 3 body parts and (A-B) did not satisfy the threshold while (B-C) did, then only C is returned. In the current implementation (B-C) is returned and (A) is excluded